Fix bug, spectrogram input tensor should have float type

+++ utils/loaders.py
@@ -492,7 +492,7 @@ class ActionNetDataset(data.Dataset, ABC):
                 'right': record.myo_right_readings.reshape(8, -1)
             }
             
-            process_data = torch.from_numpy(np.array([readings[arm][i] for arm in readings.keys() for i in range(len(readings[arm]))]))
+            process_data = torch.from_numpy(np.array([readings[arm][i] for arm in readings.keys() for i in range(len(readings[arm]))])).float()
             #logger.info(f'yo1!: {process_data.shape}')
             #process_data = readings
             if self.transform is not None:
